webapi/tct-animationtiming-w3c-tests/animationtiming/AnimationTiming_Callback_FrameRequestCallback_validTimestamp.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
         Yu, XiaoyanX <xiaoyanx.yu@intel.com>
         Cui, Jieqiong <jieqiongx.cui@intel.com>

Revision history:
Date                        Author                            Description
08-29-2012         Yu, XiaoyanX <xiaoyanx.yu@intel.com>      case creation

-->

<html>
  <head>
    <meta content="text/html; charset=utf-8" http-equiv="Content-Type" />
    <title>Check whether FrameRequestCallback contains a valid timestamp</title>
    <script type="text/javascript" src="../resources/testharness.js"></script>
    <script type="text/javascript" src="../resources/testharnessreport.js"></script>
    <link rel="stylesheet" href="../resources/testharness.css" />
  </head>
  <body>
    <div id="log"></div>
    <script>
        setup({timeout:500});
        var t = async_test("Check whether FrameRequestCallback contains a valid timestamp");
        var reqAnimAPI =   window.requestAnimationFrame || window.webkitRequestAnimationFrame || window.mozRequestAnimationFrame || window.msRequestAnimationFrame || oRequestAnimationFrame;
        try{
            reqAnimAPI(Callback);
        }catch(e){
            t.step(function() {
                assert_true(false,"The browser does not support requestAnimationFrame method");
            });
            t.done();
        }
        function Callback(timestamp)
        {
            //Check that requestAnimationFrame has a timestamp
            var callbackTime = new Date(timestamp);
            var msg = "FrameRequestCallback contains a valid timestamp: " + callbackTime.toTimeString();

            t.step(function() {
                assert_not_equals(callbackTime.valueOf(), NaN, msg);
            });
            t.done();
        }
    </script>
  </body>
</html>
